// src/interfaces/empleados.d.ts
export interface Employee {
  id: string;
  nombres: string;
  apellidos: string;
  tipo_documento: string;
  num_documento: string;
  genero: string;
  fecha_nacimiento: string;
  celular: number;
  email: string;
  direccion_domicilio: string;
  foto_url: string;
  foto_id: string;
  estado: string;
  rol: string;
  terminal_id: string;
  createdAt: string;
  updatedAt: string;
}